backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
gunicorn
flask
flask_cors
faiss-cpu
openai